refactor(core/validation): move FormValidator classes to separate files
Create a file for FormValidator and FormValidatorField.
Keep interfaces and helpers in validation.ts.
