+++ app/scripts/modules/core/src/presentation/forms/validation/index.ts
@@ -1,2 +1,4 @@
 export * from './validators';
 export * from './validation';
+export * from './FormValidator';
+export * from './FormValidatorField';
